pyproject.toml
[tool.poetry]
name = "motion-capture"
version = "0.1.0"
description = "motion capture suite"
authors = ["kristina-aino <kristina.ai@proton.me>"]
license = "MIT"
readme = "README.md"
packages = [{include = "motion_capture"}]

[tool.poetry.dependencies]
python = "^3.10"
device-capture-system = {path = "../device-capture-system/dist/device_capture_system-1.1.0.tar.gz"}
calibrate-camera-system = {path = "../calibrate-camera-system/dist/calibrate_camera_system-1.0.0.tar.gz"}
torch = "2.3.1+cu121"
torchvision = "0.18.1+cu121"
torchaudio = "2.3.1+cu121"
numpy-quaternion = "^2023.0.3"
numpy = "^1.1.0"
matplotlib = "^3.9.0"
ipython = "^8.25.0"
notebook = "^7.2.1"
pandas = "^2.2.2"
h5py = "^3.11.0"
lovely-tensors = "^0.1.15"
tensorboardx = "^2.6.2.2"
pytorch-lightning = "^2.3.1"
sympy = "^1.12.1"
tensorboard = "^2.17.0"
torch-tb-profiler = "^0.4.3"
optuna = "^3.6.1"
hydra-core = "^1.3.2"
optuna-integration = "^3.6.0"
mlflow = "^2.14.2"
lightning = "^2.3.2"
timm = "^1.0.7"
ultralytics = "^8.2.77"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
